pyproject.toml
[project]
name = "storesync"
version = "0.1.0"
description = "Multi-location marketing operations platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=24.0",
    "django-fsm>=2.8",
    "django-import-export>=3.3",
    "drf-spectacular>=0.26",
    "psycopg[binary]>=3.1",
    "pgvector>=0.2",
    "celery>=5.3",
    "redis>=5.0",
    "langchain>=0.1",
    "langchain-openai>=0.0.5",
    "openai>=1.0",
    "gunicorn>=21.0",
    "python-decouple>=3.8",
    "Pillow>=10.0",
    "Jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "pytest-cov>=4.1",
    "factory-boy>=3.3",
    "ruff>=0.1",
    "ipython>=8.0",
    "django-debug-toolbar>=4.2",
]
production = [
    "sentry-sdk>=1.38",
    "django-storages>=1.14",
    "boto3>=1.34",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "DJ", "PIE", "T20", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["apps"]
omit = ["*/migrations/*", "*/tests/*"]
